Tweaked ssh fix in profile.py
Added an explicit definition of my account for the ssh known_hosts. If this doesn't work, a static solution would probably work fine. I have thought of making a python file that also takes care of things, eventually making a .sh file that can be auto run to map all the things.